Translate C lexer tests to C unit tests
This allows us to get rid of a bunch of exports and ctypes wrappers.
