+++ tests/test_language_c.py
@@ -1,7 +1,6 @@
 # Copyright (c) Meta Platforms, Inc. and affiliates.
 # SPDX-License-Identifier: LGPL-2.1-or-later
 from functools import reduce
-from itertools import chain
 import operator
 
 from drgn import (
@@ -14,8 +13,7 @@ from drgn import (
     cast,
     container_of,
 )
-from tests import MockProgramTestCase, TestCase
-from tests.libdrgn import C_TOKEN, Lexer, drgn_c_family_lexer_func
+from tests import MockProgramTestCase
 
 
 class TestPrettyPrintTypeName(MockProgramTestCase):
@@ -709,169 +707,6 @@ typedef struct {
                 ),
             ),
         )
-
-
-class TestLexer(TestCase):
-    def lex(self, s, cpp=False):
-        lexer = Lexer(drgn_c_family_lexer_func, s, cpp)
-        while True:
-            token = lexer.pop()
-            if token.kind == C_TOKEN.EOF:
-                break
-            yield token
-
-    def test_empty(self):
-        lexer = Lexer(drgn_c_family_lexer_func, "")
-        for i in range(64):
-            self.assertEqual(lexer.pop().kind, C_TOKEN.EOF)
-
-    def test_symbols(self):
-        s = "()[]*."
-        tokens = [
-            C_TOKEN.LPAREN,
-            C_TOKEN.RPAREN,
-            C_TOKEN.LBRACKET,
-            C_TOKEN.RBRACKET,
-            C_TOKEN.ASTERISK,
-            C_TOKEN.DOT,
-        ]
-        self.assertEqual([token.kind for token in self.lex(s)], tokens)
-
-    def test_keywords(self):
-        s = """void char short int long signed unsigned _Bool float double
-        _Complex const restrict volatile _Atomic struct union class enum"""
-        tokens = [
-            C_TOKEN.VOID,
-            C_TOKEN.CHAR,
-            C_TOKEN.SHORT,
-            C_TOKEN.INT,
-            C_TOKEN.LONG,
-            C_TOKEN.SIGNED,
-            C_TOKEN.UNSIGNED,
-            C_TOKEN.BOOL,
-            C_TOKEN.FLOAT,
-            C_TOKEN.DOUBLE,
-            C_TOKEN.COMPLEX,
-            C_TOKEN.CONST,
-            C_TOKEN.RESTRICT,
-            C_TOKEN.VOLATILE,
-            C_TOKEN.ATOMIC,
-            C_TOKEN.STRUCT,
-            C_TOKEN.UNION,
-            C_TOKEN.IDENTIFIER,
-            C_TOKEN.ENUM,
-        ]
-        self.assertEqual([token.kind for token in self.lex(s)], tokens)
-
-    def test_cpp_keywords(self):
-        s = """void char short int long signed unsigned _Bool float double
-        _Complex const restrict volatile _Atomic struct union class enum"""
-        tokens = [
-            C_TOKEN.VOID,
-            C_TOKEN.CHAR,
-            C_TOKEN.SHORT,
-            C_TOKEN.INT,
-            C_TOKEN.LONG,
-            C_TOKEN.SIGNED,
-            C_TOKEN.UNSIGNED,
-            C_TOKEN.BOOL,
-            C_TOKEN.FLOAT,
-            C_TOKEN.DOUBLE,
-            C_TOKEN.COMPLEX,
-            C_TOKEN.CONST,
-            C_TOKEN.RESTRICT,
-            C_TOKEN.VOLATILE,
-            C_TOKEN.ATOMIC,
-            C_TOKEN.STRUCT,
-            C_TOKEN.UNION,
-            C_TOKEN.CLASS,
-            C_TOKEN.ENUM,
-        ]
-        self.assertEqual([token.kind for token in self.lex(s, cpp=True)], tokens)
-
-    def test_cpp_identifiers_with_template_parameters(self):
-        token_pairs = (
-            ("vector", "<int>"),
-            ("pair", "<int, double>"),
-            ("unordered_map", "<std::string, std::vector<pair<short, bool>>>"),
-            ("IntLiteral", "<123>"),
-            ("IntLiteralWithSuffix", "<123UL>"),
-            ("FloatLiteral", "<1.987>"),
-            ("FloatLiteralWithExponent", "<1.23423e+1f>"),
-            ("PointerLiteral", "<&asdf>"),
-            ("ParenthesizedPointerLiteral", "<(&asdf)>"),
-            ("CharLiteral", "<'a'>"),
-            ("CharLiteralEdgeCase1", "<'<'>"),
-            ("CharLiteralEdgeCase2", "<'>'>"),
-            ("CharLiteralEdgeCase3", r"<'\''>"),
-        )
-        self.assertEqual(
-            [
-                (token.kind, token.value)
-                for token in self.lex(
-                    " ".join("".join(pair) for pair in token_pairs), cpp=True
-                )
-            ],
-            [
-                (C_TOKEN.TEMPLATE_ARGUMENTS if i % 2 else C_TOKEN.IDENTIFIER, value)
-                for i, value in enumerate(chain.from_iterable(token_pairs))
-            ],
-        )
-
-    def test_cpp_identifiers_with_invalid_template_parameters(self):
-        for s in [
-            "vector<int",
-            "pair<<int, double>",
-            "unordered_map<string, vector<pair<short, bool>>",
-        ]:
-            self.assertRaisesRegex(
-                SyntaxError,
-                "invalid template arguments",
-                list,
-                self.lex(s, cpp=True),
-            )
-        for s in [
-            "vectorint>",
-            "pair<int, double>>",
-        ]:
-            self.assertRaisesRegex(
-                SyntaxError,
-                "invalid character",
-                list,
-                self.lex(s, cpp=True),
-            )
-
-    def test_identifiers(self):
-        s = "_ x foo _bar baz1"
-        tokens = s.split()
-        self.assertEqual(
-            [(token.kind, token.value) for token in self.lex(s)],
-            [(C_TOKEN.IDENTIFIER, value) for value in tokens],
-        )
-
-    def test_almost_keywords(self):
-        s = """voi cha shor in lon signe unsigne _Boo floa doubl
-        _Comple cons restric volatil _Atomi struc unio enu"""
-        tokens = s.split()
-        self.assertEqual(
-            [(token.kind, token.value) for token in self.lex(s)],
-            [(C_TOKEN.IDENTIFIER, value) for value in tokens],
-        )
-
-    def test_number(self):
-        s = "0 1234 0xdeadbeef"
-        tokens = s.split()
-        self.assertEqual(
-            [(token.kind, token.value) for token in self.lex(s)],
-            [(C_TOKEN.NUMBER, value) for value in tokens],
-        )
-
-    def test_invalid_number(self):
-        for s in ["0x", "1234y"]:
-            self.assertRaisesRegex(SyntaxError, "invalid number", list, self.lex(s))
-
-    def test_invalid_character(self):
-        self.assertRaisesRegex(SyntaxError, "invalid character", list, self.lex("@"))
 
 
 class TestLiteral(MockProgramTestCase):
